libs/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# Azure
azure-data-tables
azure-functions
azure-functions-durable
azure-identity
azure-keyvault-secrets
azure-storage-blob # Use datalake when possible
azure-storage-file-datalake
azure-storage-queue

# Database
pyodbc
salesforce-bulk==2.2.0
sqlalchemy

# Data Manipulation
geojson
pandas
pyarrow

# Flask
Flask
Flask-SQLAlchemy
Flask-Session
Flask-SQLAlchemy-Session
Flask-Restless-NG

# Identity/Authenticity/Authorization
py-abac
py-abac[sql]
pyjwt
requests-auth-aws-sigv4

# Other
area
numpy
pyproj
shapely
requests